GeometryLearning

Documentation for GeometryLearning.

GeometryLearning.Atten

GeometryLearning.Diag

GeometryLearning.OperatorConv

GeometryLearning.OperatorKernel

GeometryLearning.callback

GeometryLearning.model_setup

GeometryLearning.mse

GeometryLearning.optimize

GeometryLearning.rsquare

GeometryLearning.train_model

GeometryLearning.visualize

GeometryLearning.Atten — Type

Attention Layer

single layer model with no nonlinearity (single head linear attention)

u = NN(f) q = Wq * u k = Wk * u v = Wv * u

v = activation(q * k') * u

GeometryLearning.Diag — Type

Diagonal layer

GeometryLearning.OperatorConv — Type

Neural Operator convolution layer

accept data in shape (C, X1, ..., Xd, B)

GeometryLearning.OperatorKernel — Method

OperatorKernel(
    in_dims,
    out_dims,
    modes;
    activation,
    transform,
    init
)

GeometryLearning.callback — Method

callback(
    p,
    st;
    io,
    _loss,
    _LOSS,
    _stats,
    loss_,
    LOSS_,
    stats_,
    iter,
    step,
    maxiter,
    ITER
)

GeometryLearning.model_setup — Method

model_setup(NN, data)

GeometryLearning.mse — Method

mse(ypred, ytrue)

Mean squared error

GeometryLearning.optimize — Method

optimize(loss, p, st, maxiter; opt, cb)

Train parameters p to minimize loss using optimization strategy opt.

Arguments

Loss signature: loss(p, st) -> y, st

Callback signature: cb(p, st iter, maxiter) -> nothing

GeometryLearning.rsquare — Method

rsquare(ypred, ytrue) -> 1 - MSE(ytrue, ypred) / var(yture)

Calculuate r2 (coefficient of determination) score.

GeometryLearning.train_model — Method

train_model(
    rng,
    NN,
    _data,
    data_,
    V;
    opts,
    maxiters,
    cbstep,
    dir,
    name,
    io
)

Arguments

V: function space

NN: Lux neural network

_data: training data as (x, y)

data_: testing data as (x, y)

Data arrays, x, y must be AbstractMatrix, or AbstractArray{T,3}. In the former case, the dimensions are assumed to be (points, batch), and (chs, points, batch) in the latter, where the points dimension is equal to length(V).

Keyword Arguments

opts: NTuple of optimizers

maxiters: Number of iterations for each optimization cycle

cbstep: prompt callback function every cbstep epochs

dir: directory to save model, plots

io: io for printing stats

GeometryLearning.visualize — Method

visualize(V, _data, data_, NN, p, st; nsamples)