GeometryLearning

Documentation for GeometryLearning.

GeometryLearning.Atten

GeometryLearning.CosineTransform

GeometryLearning.Diag

GeometryLearning.FourierTransform

GeometryLearning.OpConv

GeometryLearning.OpConvBilinear

GeometryLearning.OpConvBilinear

GeometryLearning.OpKernel

GeometryLearning.__opconv

GeometryLearning.callback

GeometryLearning.linear_nonlinear_model

GeometryLearning.model_setup

GeometryLearning.mse

GeometryLearning.opconv__

GeometryLearning.opconv_wt

GeometryLearning.optimize

GeometryLearning.rsquare

GeometryLearning.train_model

GeometryLearning.visualize

GeometryLearning.Atten — Type

Attention Layer

single layer model with no nonlinearity (single head linear attention)

u = NN(f) q = Wq * u k = Wk * u v = Wv * u

v = activation(q * k') * u

GeometryLearning.CosineTransform — Type

struct CosineTransform{D} <: GeometryLearning.AbstractTransform{D}

GeometryLearning.Diag — Type

Diagonal layer

GeometryLearning.FourierTransform — Type

struct FourierTransform{D} <: GeometryLearning.AbstractTransform{D}

GeometryLearning.OpConv — Type

Neural Operator convolution layer

GeometryLearning.OpConvBilinear — Type

Neural Operator bilinear convolution layer

GeometryLearning.OpConvBilinear — Method

Extend OpConv to accept two inputs

Like Lux.Bilinear in modal space

GeometryLearning.OpKernel — Method

OpKernel(ch_in, ch_out, modes; activation, transform, init)

accept data in shape (C, X1, ..., Xd, B)

GeometryLearning.__opconv — Method

__opconv(x, transform, modes)

Accepts x [C, N1...Nd, B]. Returns x̂ [C, M, B] where M = prod(modes)

Operations

apply transform to N1...Nd: [K1...Kd, C, B] <- [K1...Kd, C, B]

truncate (discard high-freq modes): [M1...Md, C, B] <- [K1...Kd, C, B] where modes == (M1...Md)

GeometryLearning.callback — Method

callback(
    p,
    st;
    io,
    _loss,
    _LOSS,
    _stats,
    loss_,
    LOSS_,
    stats_,
    iter,
    step,
    maxiter,
    ITER
)

GeometryLearning.linear_nonlinear_model — Method

linear_nonlinear_model(linear, nonlinear, bilinear, project)

if you have linear dependence on x1, and nonlinear on x2, then

x1 → linear → y1 ↘
                  bilinear → project → z
x2 → nonlin → y2 ↗

GeometryLearning.model_setup — Method

model_setup(NN, data)

GeometryLearning.mse — Method

mse(ypred, ytrue)

Mean squared error

GeometryLearning.opconv__ — Method

opconv__(ŷ_tr, transform, modes, Ks, Ns)

GeometryLearning.opconv_wt — Method

opconv_wt(x, W)

Apply pointwise linear transform in mode space, i.e. no mode-mixing. Unique linear transform for each mode.

Operations

reshape: [Ci, M, B] <- [Ci, M1...Md, B] where M = prod(M1...Md)

apply weight:

reshape: [Co, M1...Md, B] <- [Co, M, B]

GeometryLearning.optimize — Method

optimize(loss, p, st, maxiter; opt, cb)

Train parameters p to minimize loss using optimization strategy opt.

Arguments

Loss signature: loss(p, st) -> y, st

Callback signature: cb(p, st iter, maxiter) -> nothing

GeometryLearning.rsquare — Method

rsquare(ypred, ytrue) -> 1 - MSE(ytrue, ypred) / var(yture)

Calculuate r2 (coefficient of determination) score.

GeometryLearning.train_model — Method

train_model(
    rng,
    NN,
    _data,
    data_,
    V;
    opts,
    maxiters,
    cbstep,
    dir,
    name,
    io
)

Arguments

V: function space

NN: Lux neural network

_data: training data as (x, y)

data_: testing data as (x, y)

Data arrays, x, y must be AbstractMatrix, or AbstractArray{T,3}. In the former case, the dimensions are assumed to be (points, batch), and (chs, points, batch) in the latter, where the points dimension is equal to length(V).

Keyword Arguments

opts: NTuple of optimizers

maxiters: Number of iterations for each optimization cycle

cbstep: prompt callback function every cbstep epochs

dir: directory to save model, plots

io: io for printing stats

GeometryLearning.visualize — Method

visualize(V, _data, data_, NN, p, st; nsamples)